app/requirements.txt
flask==2.2.3
werkzeug==2.2.3
kafka-python==2.0.2
pymongo==3.12.0
redis==4.5.1
pandas==1.5.3
numpy==1.23.5
ta-lib==0.4.24
backtrader==1.9.76.123
plotly==5.13.0
dash==2.8.1
flask-login==0.6.2
bcrypt==4.0.1